displayed restaurants, remove folder 'materials' and add folder 'constants'
